+++ types/index.d.ts
@@ -1,0 +1,7 @@
+export * from './components/components';
+export * from './components/schema';
+export * from './form';
+export * from './formbuilder';
+export * from './formio';
+export * from './templates';
+export * from './utils';